refactor: rename Executor class to NltkMeGuiExecutor
